A beautiful wavelength slider with the right ranges and the extra labels.
