+++ bending-light/bower.json
@@ -24,7 +24,6 @@
     "mocha": "~1.21.4",
     "chai": "~1.9.1",
     "nouislider": "~7.0.3 >=7.0.3",
-    "buzz": "~1.1.7",
     "pixi": "~3.0.7"
   }
 }
