+++ photoelectric-effect/bower.json
@@ -23,7 +23,7 @@
     "sinonjs": "~1.10.2",
     "mocha": "~1.21.4",
     "chai": "~1.9.1",
-    "nouislider": "~8.2.1",
+    "nouislider": "~7.0.3 >=7.0.3",
     "pixi": "~3.0.7",
     "buzz": "~1.1.7"
   }
