+++ photoelectric-effect/src/js/config.js
@@ -9,15 +9,13 @@
             bootstrap:  '../../bower_components/bootstrap/dist/js/bootstrap.min',
             text:       '../../bower_components/requirejs-text/text',
             pixi:       '../../bower_components/pixi/bin/pixi',
-            nouislider: '../../bower_components/nouislider/distribute/nouislider.min',
-            buzz:       '../../bower_components/buzz/dist/buzz.min',
+            nouislider: '../../bower_components/nouislider/distribute/jquery.nouislider.all.min',
 
             'vector2-node':          '../../node_modules/vector2-node-shimmed/index',
             'object-pool':           '../../node_modules/object-pool-shimmed/index',
             'circular-list':         '../../node_modules/object-pool-shimmed/node_modules/circular-list/index',
             'bootstrap-select':      '../../node_modules/bootstrap-select/js/bootstrap-select',
             'bootstrap-select-less': '../../node_modules/bootstrap-select/less/bootstrap-select',
-            'wnumb':                 '../../node_modules/wnumb/wNumb',
             'gauss-random':          '../../node_modules/gauss-random-shimmed/sample',
 
             'lasers':          '../../../lasers/src/js',
