+++ photoelectric-effect/src/templates/sim.html
@@ -1,24 +1,19 @@
 
 <span class="scene-view-placeholder"></span>
 
-<div class="beam-controls">
-    <div class="voltageSliderContainer">
-        <div class="voltageSlider"></div>
-        <span class="voltageSliderLabel"></span>
+<div class="sim-controls beam-controls">
+    <div class="wavelength-slider-container">
+        <label class="slider-header">Wavelength <span class="slider-value wavelength-value">400nm</span></label>
+        <div class="wavelength-slider-wrapper">
+            <div class="uv-label">UV</div>
+            <div class="ir-label">IR</div>
+        </div>
     </div>
 
-    <div class="wavelengthSliderContainer">
-        <div class="wavelengthSlider noUi-extended"></div>
-        <span class="wavelengthSliderLabel"></span>
+    <div class="intensity-slider-container">
+        <div class="intensity-slider"></div>
+        <span class="intensity-value"></span>
     </div>
-
-    <div class="intensitySliderContainer">
-        <div class="intensitySlider noUi-extended"></div>
-        <span class="intensitySliderLabel"></span>
-    </div>
-
-    <div class="uvLabel">UV</div>
-    <div class="irLabel">IR</div>
 </div>
 
 <div class="sim-controls-wrapper">
